requirements.txt
asgiref==3.5.2
channels==4.0.0
Django==4.1.4
django-compat==1.0.15
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.4
Pillow==9.3.0
prometheus-client==0.19.0
PyJWT==2.6.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
